requirements.txt
# =======================
# AgroAI Bot v2.0 - Python Dependencies
# Python 3.12.7
# Author: Sariboy Eshbayev (MYSMART Team)
# =======================

# ----------------------------------------
# Telegram Bot Framework
# ----------------------------------------
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# ----------------------------------------
# Async / Networking
# ----------------------------------------
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0

# ----------------------------------------
# Image Processing & Math
# ----------------------------------------
Pillow==10.1.0
numpy==1.26.2

# ----------------------------------------
# Planetary Computer (Real Sentinel-2 NDVI)
# IMPORTANT: API keys are NOT required.
# ----------------------------------------
planetary-computer==1.0.0
pystac-client==0.7.5
pystac==1.9.0
rasterio==1.3.9
rioxarray==0.15.0
xarray==2023.12.0
shapely==2.0.2

# ----------------------------------------
# AI Models (Anthropic Claude)
# ----------------------------------------
anthropic==0.8.1

# ----------------------------------------
# Configuration & Utils
# ----------------------------------------
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==2.5.2
colorlog==6.8.0

# ----------------------------------------
# Testing
# ----------------------------------------
pytest==7.4.3
pytest-asyncio==0.21.1
